pyproject.toml
[tool.poetry]
name = "quiz"
version = "0.1.0"
description = "a quiz app that i was forced to make"
authors = ["ezntek <eason@ezntek.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
colorama = "^0.4.6"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.1"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
split_before_logical_operator = true
blank_lines_around_top_level_definition = 1
blank_line_before_nested_class_or_def = false
coalesce_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
